fixes and new builtins
- `str::split`

fixed:
- codegen around nested blocks
- iterating call frames during unwinding was done in the wrong order

i previously made the assumption that variables are always at the bottom
of the stack for a given function, but that is trivially not true:

```
1 + do { let v = 0 }
```

this has a temporary at the bottom of the stack (the `1`). so the fix
was to remove this assumption. this actually ended up making codegen
better in a few other areas, too.

we iterate through call frames on error to find the first script call
frame, and use its debug info to report the error at a suitable span. we
cant really report spans from within host functions, so this is the next
best thing. the issue was that we were iterating from the bottom-most
call frame up, which meant we'd always immediately arrive at @main. now
we iterate from the top-most call frame down, which does the right
thing.